Cleanup duplicate key impls
