+++ src/duplicate_key_impls/mod.rs
@@ -1,0 +1,5 @@
+mod error_on_duplicate;
+
+pub use self::{
+    error_on_duplicate::{PreventDuplicateInsertsMap, PreventDuplicateInsertsSet},
+};
